refactor : 코드 정리, dto 재생성
