+++ src/main/java/com/my/relink/service/DonationItemService.java
@@ -1,17 +1,16 @@
 package com.my.relink.service;
 
 import com.my.relink.config.security.AuthUser;
+import com.my.relink.controller.item.donation.dto.DonationItemReqDto;
 import com.my.relink.domain.category.Category;
+import com.my.relink.domain.category.repository.CategoryRepository;
 import com.my.relink.domain.item.donation.DonationItem;
-import com.my.relink.domain.item.donation.ItemQuality;
 import com.my.relink.domain.user.User;
-import com.my.relink.controller.item.donation.dto.req.DonationItemReqDto;
-import com.my.relink.controller.item.donation.dto.resp.DonationItemRespDto;
+import com.my.relink.controller.item.donation.dto.DonationItemRespDto;
+import com.my.relink.domain.user.repository.UserRepository;
 import com.my.relink.ex.BusinessException;
 import com.my.relink.ex.ErrorCode;
-import com.my.relink.controller.item.donation.repository.CategoryRepository;
 import com.my.relink.controller.item.donation.repository.DonationItemRepository;
-import com.my.relink.controller.item.donation.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
